fix webpack config
