--- src/lambda.ts
@@ -1,35 +1,0 @@
-import { NestFactory } from '@nestjs/core';
-import { ExpressAdapter } from '@nestjs/platform-express';
-import serverlessExpress from '@vendia/serverless-express';
-import { Context, Handler } from 'aws-lambda';
-import express from 'express';
-
-import { AppModule } from './app.module';
-
-let cachedServer: Handler;
-
-async function bootstrap() {
-  if (!cachedServer) {
-    const expressApp = express();
-    const nestApp = await NestFactory.create(
-      AppModule,
-      new ExpressAdapter(expressApp),
-    );
-
-    nestApp.enableCors();
-
-    await nestApp.init();
-
-    cachedServer = serverlessExpress({ app: expressApp });
-  }
-
-  return cachedServer;
-}
-
-const handler = async (event: any, context: Context, callback: any) => {
-  const server = await bootstrap();
-     console.log('Event-[%s], context-[%s]', event, context);
-  return server(event, context, callback);
-};
-
-module.exports.handler = handler;
